requirements.txt
streamlit
playwright
google-generativeai